Move Stream type into its own module
